update: search by status - 4/3/2024

+++ demo/src/main/java/demo/service/QuestionService.java
@@ -5,6 +5,7 @@ import demo.repository.QuestionRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.nio.file.*;
 import java.util.Date;
 import java.util.List;
 import java.util.Optional;
@@ -18,11 +19,25 @@ public class QuestionService{
         return questionRepository.findAll();
     }
 
-
     public void deleteQuestionById(int id) {
+        Questions question = questionRepository.findById(id).orElse(null);
+        if (question != null) {
 
-        questionRepository.deleteById(id);
+            if (question.getImage() != null) {
+                String imagePath = "public/images/" + question.getImage();
+                try {
+                    Path path = Paths.get(imagePath);
+                    Files.deleteIfExists(path);
+                } catch (Exception e) {
+
+                    e.printStackTrace();
+                }
+            }
+
+            questionRepository.deleteById(id);
+        }
     }
+
     public void save(Questions questions){
         questionRepository.save(questions);
     }
